+++ integrationtests/drop_test.go
@@ -38,7 +38,7 @@ var _ = Describe("Drop Proxy", func() {
 			"https://quic.clemente.io/data",
 		)
 
-		session, err := Start(command, GinkgoWriter, GinkgoWriter)
+		session, err := Start(command, nil, GinkgoWriter)
 		Expect(err).NotTo(HaveOccurred())
 		defer session.Kill()
 		Eventually(session, 20).Should(Exit(0))